% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dire_bonjour.R
\name{dire_bonjour}
\alias{dire_bonjour}
\title{dire_bonjour}
\usage{
dire_bonjour(prenom = NULL)
}
\arguments{
\item{prenom}{optional name of the person to whom say hello!}
}
\value{
a character string giving the gear
}
\description{
\code{dire_bonjour} say hello !
}
\examples{

dire_bonjour()

dire_bonjour("paul")

}
\author{
Jonathan Rault, \email{jonathan.rault@ifremer.fr}
}
