% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pays}
\alias{pays}
\title{pays data}
\format{
A data frame with 5 rows and 2 variables:
\describe{
\item{ISO2}{ISO2 Code}
\item{ISO3}{ISO3 Code}
\item{NAME_FRA}{French name of the country}
\item{NAME_ENG}{English name of the country}
}
}
\usage{
pays
}
\description{
a dataset containing ISO2 and ISO3 codes for all countries
}
\keyword{datasets}
